Use sys.executable for default python and remove __PYVENV_LAUNCHER__ to avoid issues with virtualenv
Using a venv on Windows adds the environment variable '__PYVENV_LAUNCHER__' which overrides subprocess calls to python and to the system python.
This prevents a command from invoking the system python, which is necessary to setup new virtual environments.
The workaround is to manually remove this environment variable and default to sys.executable instead of 'python', since python would resolve to the system python at this point.