Check for exit codes from signals in integration tests because they randomly come up
The test test_master_disconnection randomly fails because of the master process exiting with a result code from the shutdown signal.
After checking logs and debugging, the test appears to run as expected, including the shutdown behavior and it should exit with the code 0 for success.
For Linux, it could be some weird behavior with Python exit logic, but for the most part I have no idea what is happening.