Delay update processing to fix a race condition with synchronization
The messenger and worker implementation are so that the updates are processed immediately while responses are processed when the async caller resumes.
In the case where the run completes very quickly, the resynchronize response and the synchronization completion event can be sent together by a single messenger push.
On the master side, the updates are processed and the synchronization is set to done, then the resynchronize response is processed and the synchronization is set to running.
This prevents the executor from proceeding to the finishing state and from completing cleanly.